Scum runs

+++ src/edu/ucsb/cs56/projects/games/fish_animation/Bubbles.java
@@ -72,7 +72,7 @@ public class Bubbles extends ScumOfTheSea
     /**Changes the position of the Bubble
      * @param m changes the position of the Bubbles object
      */
-    @Override void move(int m){
+    @Override void move(double m){
 	y-= m;
     }
 }
